splitSCript
